Array to JSON
Renvoie un objet JSON pour FILMByID et ActeurById

+++ ProjetCinemaAPI/src/main/java/com/webservice/projetcinema/repository/ActeurRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ActeurRepository extends JpaRepository<Acteur, Long> {
+public interface ActeurRepository extends JpaRepository<Acteur, Integer> {
 
 
     @Query("Select A.NoAct,A.NomAct,A.PrenAct,A.DateNaiss,A.DateDeces" +
